♻️ [environment] => move `FlowSwitch` in dedicated module
