+++ src/youwol/app/environment/models/__init__.py
@@ -4,6 +4,7 @@ The models used in YouWolEnvironment and its Configuration.
 
 # relative
 from .defaults import *
+from .flow_switches import *
 from .models import *
 from .models_config import *
 from .predefined_configs import *
